update and refactor to match Fabbri (2008)
