+++ docs/references.md
@@ -1,7 +1,23 @@
 # References
 
 1. Griffiths, D. J. (2013) *Introduction to Electrodynamics* (4th ed.). 
-2. Thorne, R. J. & Tarrh, J. M. (1982). *MHD and Fusion Magnets: Field and Force Design Concepts*.
-3. Bottura, L. (2006, September 30). Magnum - A Fortran Library for the Calculation of Magnetic Configurations.[https://supermagnet.sourceforge.io/notes/CRYO-06-034.pdf](https://supermagnet.sourceforge.io/notes/CRYO-06-034.pdf)
-3. Field Precision, LLC. (2026, March 10). Theory and applications of the Maxwell stress tensor.[https://fieldp.com/tutorials/stresstensor.pdf](https://fieldp.com/tutorials/stresstensor.pdf)  
-4. Simpson, Lane, Immer, & Youngquist. (2001, February 1). Simple Analytic Expressions for the Magnetic Field of a Circular Current Loop. National Aeronautics and Space Administration. [https://ntrs.nasa.gov/citations/20140002333](https://ntrs.nasa.gov/citations/20140002333)
+2. Thorne, R. J. & Tarrh, J. M. (1982). 
+*MHD and Fusion Magnets: Field and Force Design Concepts*.
+3. Bottura, L. (2006, September 30). Magnum - A Fortran Library for the 
+Calculation of Magnetic Configurations.
+[https://supermagnet.sourceforge.io/notes/CRYO-06-034.pdf](https://supermagnet.sourceforge.io/notes/CRYO-06-034.pdf)
+3. Field Precision, LLC. (2026, March 10). Theory and applications of the Maxwell 
+stress tensor.
+[https://fieldp.com/tutorials/stresstensor.pdf](https://fieldp.com/tutorials/stresstensor.pdf)  
+4. Simpson, Lane, Immer, & Youngquist. (2001, February 1). 
+Simple Analytic Expressions for the Magnetic Field of a Circular Current Loop. 
+National Aeronautics and Space Administration. 
+[https://ntrs.nasa.gov/citations/20140002333](https://ntrs.nasa.gov/citations/20140002333)
+5. M. Fabbri, "Magnetic Flux Density and Vector Potential of Uniform 
+Polyhedral Sources," in IEEE Transactions on Magnetics, vol. 44, no. 1, pp. 32-36, Jan. 2008. 
+<https://ieeexplore.ieee.org/document/4407584>.
+6. Byzov, D., Martyshko, P., & Chernoskutov, A. (2022).
+Computationally Effective Modeling of Self-Demagnetization and Magnetic Field for 
+Bodies of Arbitrary Shape Using Polyhedron Discretization.
+Mathematics, 10(10), 1656.
+<https://doi.org/10.3390/math10101656>
