+++ src/sources/tet4.rs
@@ -1,444 +1,21 @@
-#![allow(non_snake_case, unused)]
 use crate::{
     math::{atan2, ln},
-    sources::element::{edge_csys, edge_integral, edge_integral_gradient, transform},
-    types::{Mat3, Vec3},
+    types::Vec3,
 };
 
 use std::f64::consts::PI;
-
-// Refer to the gmsh or ansys documentation for node numbering on a tet element:
-// https://gmsh.info/doc/texinfo/gmsh.html#Node-ordering
-// https://ansyshelp.ansys.com/public/account/secured?returnurl=//////////Views/Secured/corp/v242/en/ans_elem/Hlp_E_SOLID236.html
-const NODE_MAP: [[[usize; 3]; 3]; 4] = [
-    [[0, 2, 1], [2, 1, 0], [1, 0, 2]], // back face
-    [[0, 3, 2], [3, 2, 0], [2, 0, 3]], // left face
-    [[1, 2, 3], [2, 3, 1], [3, 1, 2]], // right face
-    [[0, 1, 3], [1, 3, 0], [3, 0, 1]], // bottom face
-];
-
 const INV_4PI: f64 = 1.0 / (4.0 * PI);
 
-/// Compute magnetic field intensity generated by a 4-node tetrahedral finite element
-///
-/// This function operates on n-number of target points and
-/// is vectorized for efficient computation. This version uses the Bottura edge-based
-/// formulation (see reference.)
-///
-/// # Arguments
-/// - `nodes`: (m), coordinates of each corner node in 3D space
-/// - `jdensity`: (A/m2), current density vector, assumed constant throughout the element
-/// - `targets`: (m),  target locations in 3d space
-/// - `f`: workspace allocated by the caller
-/// - `h`: (A/m), preallocated arrays for h-field results
-///
-/// # Reference  
-///
-/// CRYO-06-034: "MAGNUM - A Fortran Library for the Calculation of Magnetic Configurations"  
-/// [https://supermagnet.sourceforge.io/notes/CRYO-06-034.pdf](https://supermagnet.sourceforge.io/notes/CRYO-06-034.pdf)
-pub fn h_current_tet4_edge(
-    nodes: &[Vec3; 4],
-    jdensity: &Vec3,
-    targets: (&[f64], &[f64], &[f64]),
-    f: &mut [Vec3],
-    h: (&mut [f64], &mut [f64], &mut [f64]),
-) {
-    assert_eq!(f.len(), targets.0.len());
+// Refer to the gmsh or ansys documentation for node numbering on a tet element:
+// <https://gmsh.info/doc/texinfo/gmsh.html#Node-ordering>
+// <https://ansyshelp.ansys.com/public/account/secured?returnurl=//////////Views/Secured/corp/v242/en/ans_elem/Hlp_E_SOLID236.html>
+const NODE_WINDING: [(usize, usize, usize); 4] = [(0, 2, 1), (0, 3, 2), (1, 2, 3), (0, 1, 3)];
 
-    // Allocate memory for the result of the sum before crossing with J
-    // TODO: replace?
-    // let mut f: Vec<Vec3> = vec![Vec3([0.0; 3]); targets.0.len()];
-
-    for face in 0..4 {
-        for edge in 0..3 {
-            // normal on face
-
-            // Node numbers from the map
-            let n0: usize = NODE_MAP[face][edge][0];
-            let n1: usize = NODE_MAP[face][edge][1];
-            let n2: usize = NODE_MAP[face][edge][2];
-
-            // Local coordinate system tensor; transform the target point into the local system
-            let (xhat, yhat, zhat) = edge_csys(&nodes[n0], &nodes[n1], &nodes[n2]);
-            let ni: Vec3 = zhat;
-            let xpq2 = (nodes[n1] - nodes[n0]).mag(); // length of the edge
-
-            // Compute the effect of this edge by transforming each target into the local reference frame
-            for i in 0..targets.0.len() {
-                let target_local: Vec3 = Vec3([
-                    targets.0[i] - nodes[n1][0],
-                    targets.1[i] - nodes[n1][1],
-                    targets.2[i] - nodes[n1][2],
-                ]);
-                let p = transform(&target_local, &xhat, &yhat, &zhat);
-                let s_face = p[1]
-                    * (edge_integral(xpq2 - p[0], p[1], p[2]) - edge_integral(-p[0], p[1], p[2]));
-                f[i] += ni * s_face;
-            }
-        }
-    }
-
-    // Cross with J/4pi and accumulate
-    for i in 0..targets.0.len() {
-        let _h = jdensity.cross(&f[i]) * INV_4PI;
-        h.0[i] += _h[0];
-        h.1[i] += _h[1];
-        h.2[i] += _h[2];
-    }
-}
-
-// Contains data necessary for coordinate transforms and computations on a tet4 edge
-#[derive(Copy, Clone)]
-struct Edge {
-    xhat: Vec3,
-    yhat: Vec3,
-    zhat: Vec3,
-    ni: Vec3, // normal
-    origin: Vec3,
-    xpq2: f64,                 // length of edge
-    j_cross_ni_over_4pi: Vec3, // for the accumulator, precomputed
-}
-
-impl Default for Edge {
-    fn default() -> Self {
-        Self {
-            xhat: Vec3::default(),
-            yhat: Vec3::default(),
-            zhat: Vec3::default(),
-            ni: Vec3::default(),
-            origin: Vec3::default(),
-            xpq2: 0.0,
-            j_cross_ni_over_4pi: Vec3::default(),
-        }
-    }
-}
-
-// Compute edge data for 12x edges on the element (3x per 4x faces)
-fn compute_edges(nodes: &[Vec3; 4], jdensity: &Vec3) -> [Edge; 12] {
-    let mut edges: [Edge; 12] = [Edge::default(); 12];
-    for f in 0..4 {
-        for e in 0..3 {
-            let idx: usize = f * 3 + e;
-            // Node numbers from the map
-            let n0: usize = NODE_MAP[f][e][0];
-            let n1: usize = NODE_MAP[f][e][1];
-            let n2: usize = NODE_MAP[f][e][2];
-
-            // Local coordinate system tensor; transform the target point into the local system
-            let (xhat, yhat, zhat) = edge_csys(&nodes[n0], &nodes[n1], &nodes[n2]);
-            let ni: Vec3 = zhat;
-            let xpq2 = (nodes[n1] - nodes[n0]).mag(); // length of the edge
-
-            // Assignment
-            edges[idx].xhat = xhat;
-            edges[idx].yhat = yhat;
-            edges[idx].zhat = zhat;
-            edges[idx].ni = ni;
-            edges[idx].origin = nodes[n1];
-            edges[idx].xpq2 = xpq2;
-            edges[idx].j_cross_ni_over_4pi = jdensity.cross(&ni) * INV_4PI;
-        }
-    }
-    edges
-}
-
-// Formulation of the above that has a simpler structure and reuses edge data
-// Does not auto-vectorize and is generally 30-40% slower
-//
-// TODO: vectorize inner loop to make this cleaner and faster than the existing
-// implementation
-fn _h_field_tet4(
-    nodes: &[Vec3; 4],
-    jdensity: &Vec3,
-    targets: (&[f64], &[f64], &[f64]),
-    out: (&mut [f64], &mut [f64], &mut [f64]),
-) {
-    // Length checks
-    let (x, y, z) = targets;
-    let (hx, hy, hz) = out;
-    let n_targets = x.len();
-    assert_eq!(n_targets, y.len());
-    assert_eq!(n_targets, z.len());
-    assert_eq!(n_targets, hx.len());
-    assert_eq!(n_targets, hy.len());
-    assert_eq!(n_targets, hz.len());
-
-    // Precompute the edge data for all edges of this element
-    let edges: [Edge; 12] = compute_edges(nodes, jdensity);
-
-    // For each target point, compute all contributions from all edges
-    for i in 0..n_targets {
-        let target: Vec3 = Vec3([x[i], y[i], z[i]]);
-        // let mut f: Vec3 = Vec3::default();    // Accumulator for this target
-
-        for edge in edges {
-            let target_local: Vec3 = target - edge.origin;
-            let p: Vec3 = transform(&target_local, &edge.xhat, &edge.yhat, &edge.zhat);
-            let s_face: f64 = p[1]
-                * (edge_integral(edge.xpq2 - p[0], p[1], p[2]) - edge_integral(-p[0], p[1], p[2]));
-            // f += edge.ni * s_face;
-            hx[i] += edge.j_cross_ni_over_4pi[0] * s_face;
-            hy[i] += edge.j_cross_ni_over_4pi[1] * s_face;
-            hz[i] += edge.j_cross_ni_over_4pi[2] * s_face;
-        }
-
-        // let h = jdensity.cross(&f) * INV_4PI;
-        // hx[i] += h[0];
-        // hy[i] += h[1];
-        // hz[i] += h[2];
-    }
-}
-
-/// Compute magnetic field intensity generated by a 4-node tetrahedral finite element
-///
-/// This function operates on n-number of target points and
-/// is vectorized for efficient computation. This version uses the solid angle
-/// formulation (see reference).
-///
-/// # Arguments
-/// - `nodes`: (m), coordinates of each corner node in 3D space
-/// - `jdensity`: (A/m2), current density vector, assumed constant throughout the element
-/// - `targets`: (m),  target locations in 3d space
-/// - `h`: (A/m), preallocated arrays for h-field results
-///
-/// # Reference  
-/// Byzov, D., Martyshko, P., & Chernoskutov, A. (2022).
-/// Computationally Effective Modeling of Self-Demagnetization and Magnetic Field
-/// for Bodies of Arbitrary Shape Using Polyhedron Discretization.
-/// Mathematics, 10(10), 1656. 
-/// <https://doi.org/10.3390/math10101656>
-pub fn h_current_tet4(
-    nodes: &[Vec3; 4],
-    jdensity: &Vec3,
-    targets: (&[f64], &[f64], &[f64]),
-    h: (&mut [f64], &mut [f64], &mut [f64]),
-) {
-    let (x, y, z) = targets;
-    let (hx, hy, hz) = h;
-    let n_targets = x.len();
-    assert_eq!(n_targets, y.len());
-    assert_eq!(n_targets, z.len());
-    assert_eq!(n_targets, hx.len());
-    assert_eq!(n_targets, hy.len());
-    assert_eq!(n_targets, hz.len());
-
-    let scalar = 1.0 / (4.0 * PI);
-
-    for f in 0..4 {
-        let na = NODE_MAP[f][0][0];
-        let nb = NODE_MAP[f][0][1];
-        let nc = NODE_MAP[f][0][2];
-
-        let A: Vec3 = nodes[na];
-        let B: Vec3 = nodes[nb];
-        let C: Vec3 = nodes[nc];
-
-        let face = precompute_face(&A, &B, &C);
-        let prefactor = jdensity.cross(&face.n_hat) * scalar;
-
-        for i in 0..n_targets {
-            let r = Vec3([x[i], y[i], z[i]]);
-            let psi = charge_potential(&face, &r);
-            hx[i] += prefactor[0] * psi;
-            hy[i] += prefactor[1] * psi;
-            hz[i] += prefactor[2] * psi;
-        }
-    }
-}
-
-/// DEPRECATED
-///
-/// Compute the h field generated by a single magnetized tetrahedral element at the centroid
-/// of each element in a target mesh
-///
-/// This computes equations 58 and 60 in CRYO-06-034, assuming the target point is outside the element:  
-/// $$ H(\vec{r}) = \frac{-M}{4\pi} \cdot \nabla \int_V \frac{\vec{r}}{r^3}dV $$
-///
-/// # Arguments
-/// * `nodes`: (m) coordinates of the corner nodes of the source element
-/// * `mvector`: (A/m) magnetization vector, assumed constant over the element
-/// * `j_invt`: (1/m) inverse transpose of the Jacobian matrix as defined in `crate::gradient`
-/// * `target_nodes`: (m) x,y,z coordinates of the elements in the target mesh
-/// * `target_element_connectivity`: node indices for each element in the target mesh
-/// * `f`: pre-allocated scratch workspace of length `n_target_nodes`
-/// * `h`: (x,y,z) pre-allocated results accumulation for h-field
-pub fn hmag_tet4_grad(
-    nodes: &[Vec3; 4], // source nodes
-    mvector: &Vec3,    // source element
-    j_invt: &[Mat3],   // at each *target* element
-    target_nodes: (&[f64], &[f64], &[f64]),
-    target_element_connectivity: &[[u32; 4]],
-    f: &mut [Vec3],
-    h: (&mut [f64], &mut [f64], &mut [f64]),
-) {
-    // Temporary storage is for nodes, hmag field is calculated at elements
-    assert_eq!(f.len(), target_nodes.0.len());
-    assert_eq!(h.0.len(), target_element_connectivity.len());
-
-    // Ensure f is zeroed
-    for v in f.iter_mut() {
-        *v = Vec3::default();
-    }
-
-    // Compute the value of the all edge integrals at each target node
-    for face in 0..4 {
-        for edge in 0..3 {
-            // Node numbers from the map
-            let n0: usize = NODE_MAP[face][edge][0];
-            let n1: usize = NODE_MAP[face][edge][1];
-            let n2: usize = NODE_MAP[face][edge][2];
-
-            // Local coordinate system tensor; transform the target point into the local system
-            let (xhat, yhat, zhat) = edge_csys(&nodes[n0], &nodes[n1], &nodes[n2]);
-            let ni: Vec3 = zhat; // normal on face 
-            let xpq2 = (nodes[n1] - nodes[n0]).mag(); // length of the edge
-
-            // Compute the effect of this edge by transforming each target into the local reference frame
-            for i in 0..target_nodes.0.len() {
-                let target_local: Vec3 = Vec3([
-                    target_nodes.0[i] - nodes[n1][0],
-                    target_nodes.1[i] - nodes[n1][1],
-                    target_nodes.2[i] - nodes[n1][2],
-                ]);
-                println!("nodal target: {:?}", target_local);
-                let p = transform(&target_local, &xhat, &yhat, &zhat);
-                let s_face = p[1]
-                    * (edge_integral(xpq2 - p[0], p[1], p[2]) - edge_integral(-p[0], p[1], p[2]));
-                f[i] += ni * s_face;
-            }
-        }
-    }
-
-    // Take gradient at each element, dot with -M/4pi, and accumulate
-    let scalar: f64 = -1.0 / (4.0 * PI);
-    let m_4pi: Vec3 = *mvector * scalar;
-    let n_elements = target_element_connectivity.len();
-    for i in 0..n_elements {
-        let n_idx = target_element_connectivity[i];
-
-        // Edge/surface integral at each of the four nodes
-        let i0 = f[n_idx[0] as usize];
-        let i1 = f[n_idx[1] as usize];
-        let i2 = f[n_idx[2] as usize];
-        let i3 = f[n_idx[3] as usize];
-
-        // Compute individual `g` vectors for each component of the gradient evaluation
-        let gx = Vec3([i1[0] - i0[0], i2[0] - i0[0], i3[0] - i0[0]]);
-        let gy = Vec3([i1[1] - i0[1], i2[1] - i0[1], i3[1] - i0[1]]);
-        let gz = Vec3([i1[2] - i0[2], i2[2] - i0[2], i3[2] - i0[2]]);
-
-        // Compute gradient
-        let grad_ix = j_invt[i].mul_vec(&gx);
-        let grad_iy = j_invt[i].mul_vec(&gy);
-        let grad_iz = j_invt[i].mul_vec(&gz);
-
-        let jacobian = Mat3::from_cols(&grad_ix, &grad_iy, &grad_iz);
-        let _h_mag = jacobian.mul_vec(&m_4pi);
-        h.0[i] += _h_mag[0];
-        h.1[i] += _h_mag[1];
-        h.2[i] += _h_mag[2];
-    }
-}
-
-/// Compute the H field generated by a magnetized 4-node tetrahedral element
-///
-/// This function is vectorized over target points.
-///
-/// This computes equations 58 and 60 in CRYO-06-034, assuming the target point is outside the element:  
-/// $$ H(\vec{r}) = \frac{-M}{4\pi} \cdot \nabla \int_V \frac{\vec{r}}{r^3}dV $$
-///
-/// # Arguments
-/// * `nodes`: (m) coordinates of the corner nodes of the source element
-/// * `mvector`: (A/m) magnetization vector, assumed constant over the element
-/// * `targets`: (m) x,y,z coordinates of each target point
-/// * `h`: (x,y,z) pre-allocated results accumulation for h-field
-///
-/// # Note
-/// Accumulates into `h`. Caller must zero `h` before the first call.
-///
-/// # Reference
-/// <https://supermagnet.sourceforge.io/notes/CRYO-06-034.pdf>
-pub fn h_mag_tet4_edge(
-    nodes: &[Vec3; 4], // source nodes
-    mvector: &Vec3,    // source element
-    targets: (&[f64], &[f64], &[f64]),
-    workspace: (&mut [Vec3], &mut [Vec3], &mut [Vec3]),
-    h: (&mut [f64], &mut [f64], &mut [f64]),
-) {
-    let n_targets = targets.0.len();
-    let (dv_dx, dv_dy, dv_dz) = workspace;
-    let (hx, hy, hz) = h;
-    let scalar = -1.0 / (4.0 * PI);
-
-    // Ensure the workspace is zeroed
-    for dv in dv_dx.iter_mut() {
-        *dv = Vec3::default();
-    }
-    for dv in dv_dy.iter_mut() {
-        *dv = Vec3::default();
-    }
-    for dv in dv_dz.iter_mut() {
-        *dv = Vec3::default();
-    }
-
-    // Accumulate the effect of all element edges into the workspace
-    for face in 0..4 {
-        for edge in 0..3 {
-            // Node numbers from the map
-            let n0: usize = NODE_MAP[face][edge][0];
-            let n1: usize = NODE_MAP[face][edge][1];
-            let n2: usize = NODE_MAP[face][edge][2];
-
-            // Local coordinate system tensor; transform the target point into the local system
-            let (xhat, yhat, zhat) = edge_csys(&nodes[n0], &nodes[n1], &nodes[n2]);
-            let ni: Vec3 = zhat;
-            let xpq2 = (nodes[n1] - nodes[n0]).mag(); // length of the edge
-
-            // Compute the effect of this edge by transforming each target into the local reference frame
-            for i in 0..targets.0.len() {
-                let target_local: Vec3 = Vec3([
-                    targets.0[i] - nodes[n1][0],
-                    targets.1[i] - nodes[n1][1],
-                    targets.2[i] - nodes[n1][2],
-                ]);
-                let p = transform(&target_local, &xhat, &yhat, &zhat);
-
-                // Edge integral: E[end] - E[start]
-                let e_diff =
-                    edge_integral(xpq2 - p[0], p[1], p[2]) - edge_integral(-p[0], p[1], p[2]);
-
-                // Edge integral gradients at the start and end points
-                // Each is a Vec3 = (de_dx, de_dy, de_dz)
-                let grad_start: Vec3 = edge_integral_gradient(-p[0], p[1], p[2]);
-                let grad_end: Vec3 = edge_integral_gradient(xpq2 - p[0], p[1], p[2]);
-
-                // Chain rule
-                let ds_dx: f64 = -p[1] * (grad_end[0] - grad_start[0]);
-                let ds_dy: f64 = e_diff + p[1] * (grad_end[1] - grad_start[1]);
-                let ds_dz: f64 = p[1] * (grad_end[2] - grad_start[2]);
-
-                // Transform local gradients into global frame using basis vectors (eq 61)
-                let grad_s_x: f64 = xhat[0] * ds_dx + yhat[0] * ds_dy + zhat[0] * ds_dz;
-                let grad_s_y: f64 = xhat[1] * ds_dx + yhat[1] * ds_dy + zhat[1] * ds_dz;
-                let grad_s_z: f64 = xhat[2] * ds_dx + yhat[2] * ds_dy + zhat[2] * ds_dz;
-
-                // Accumulate into workspace
-                dv_dx[i] += ni * grad_s_x;
-                dv_dy[i] += ni * grad_s_y;
-                dv_dz[i] += ni * grad_s_z;
-            }
-        }
-    }
-
-    // Dot with M and accumulate
-    for i in 0..n_targets {
-        hx[i] += scalar
-            * (mvector[0] * dv_dx[i][0] + mvector[1] * dv_dx[i][1] + mvector[2] * dv_dx[i][2]);
-        hy[i] += scalar
-            * (mvector[0] * dv_dy[i][0] + mvector[1] * dv_dy[i][1] + mvector[2] * dv_dy[i][2]);
-        hz[i] += scalar
-            * (mvector[0] * dv_dz[i][0] + mvector[1] * dv_dz[i][1] + mvector[2] * dv_dz[i][2]);
-    }
+// Return the node indices associated with a given triangular face, using the proper
+// winding order: the perimeter of the face is traversed according to the right hand
+// rule, with the normal facing away from the center of the element.
+fn face_nodes(f: usize) -> (usize, usize, usize) {
+    NODE_WINDING[f]
 }
 
 // Per-face data used for solid angle - based calculations
@@ -462,21 +39,21 @@ struct Face {
 // Compute the unit vectors associated with a face given vertice A, B, C
 // Vertices are given in counter-clockwise direction according to the right hand
 // rule, to produce an outward facing normal vector
-fn precompute_face(A: &Vec3, B: &Vec3, C: &Vec3) -> Face {
+fn precompute_face(a: &Vec3, b: &Vec3, c: &Vec3) -> Face {
     let mut face: Face = Face {
-        a: *A,
-        b: *B,
-        c: *C,
+        a: *a,
+        b: *b,
+        c: *c,
         ..Default::default()
     };
 
     // Face unit normal
-    face.n_hat = (*B - *A).cross(&(*C - *A));
+    face.n_hat = (*b - *a).cross(&(*c - *a));
     face.n_hat /= face.n_hat.mag();
 
-    for (e, (P, Q)) in [(A, B), (B, C), (C, A)].into_iter().enumerate() {
+    for (e, (p, q)) in [(a, b), (b, c), (c, a)].into_iter().enumerate() {
         // Unit vector along edge
-        let mut e_hat = *Q - *P;
+        let mut e_hat = *q - *p;
         e_hat /= e_hat.mag();
 
         // Unit vector in plane, perpendicular to edge and normal
@@ -496,7 +73,10 @@ fn precompute_face(A: &Vec3, B: &Vec3, C: &Vec3) -> Face {
 // solid angle
 //
 // # Reference
-// <https://en.wikipedia.org/wiki/Solid_angle#Tetrahedron>
+// A. Van Oosterom and J. Strackee, "The Solid Angle of a Plane Triangle,"
+// in IEEE Transactions on Biomedical Engineering,
+// vol. BME-30, no. 2, pp. 125-126, Feb. 1983,
+// doi: <10.1109/TBME.1983.325207>.
 #[allow(non_snake_case)]
 #[inline]
 fn solid_angle(A: &Vec3, B: &Vec3, C: &Vec3, r: &Vec3) -> f64 {
@@ -517,106 +97,42 @@ fn solid_angle(A: &Vec3, B: &Vec3, C: &Vec3, r: &Vec3) -> f64 {
     2.0 * atan2(num, den)
 }
 
-// Computes solid angle and gradient of solid angle to reuse values
-#[allow(non_snake_case)]
-#[inline]
-fn solid_angle_and_gradient(A: &Vec3, B: &Vec3, C: &Vec3, r: &Vec3) -> (f64, Vec3) {
-    let a: Vec3 = *A - *r;
-    let b: Vec3 = *B - *r;
-    let c: Vec3 = *C - *r;
-    let a_mag: f64 = a.mag();
-    let b_mag: f64 = b.mag();
-    let c_mag: f64 = c.mag();
+// Edge potential as defined in eq. 18 in Fabbri (2008).
+//
+// This function is NOT regularized and therefore is singular on the edge
+fn edge_potential(r1: &Vec3, r2: &Vec3, r: &Vec3) -> f64 {
+    // Distance from r1 to r, r2 to r, and r1 to r2
+    let d1: f64 = (*r1 - *r).mag();
+    let d2: f64 = (*r2 - *r).mag();
+    let e: f64 = (*r2 - *r1).mag();
+    let s: f64 = d1 + d2;
 
-    let num: f64 = a.dot(&b.cross(&c));
-    let den1: f64 = a_mag * b_mag * c_mag;
-    let den2: f64 = a.dot(&b) * c_mag;
-    let den3: f64 = a.dot(&c) * b_mag;
-    let den4: f64 = b.dot(&c) * a_mag;
-    let den: f64 = den1 + den2 + den3 + den4;
+    // Edge potential in stable form
+    // 2.0 * ln(s + e) - ln(2.0 * (d1 + d2 + (*r1 - *r).dot(&(*r2 - *r))));
 
-    // Solid angle calculation complete
-    let angle: f64 = 2.0 * atan2(num, den);
-
-    let grad_num: Vec3 = -(b.cross(&c) + c.cross(&a) + a.cross(&b));
-
-    let s_a: f64 = (b.dot(&c) + b_mag * c_mag) / a_mag + b_mag + c_mag;
-    let s_b: f64 = (a.dot(&c) + a_mag * c_mag) / b_mag + a_mag + c_mag;
-    let s_c: f64 = (a.dot(&b) + a_mag * b_mag) / c_mag + a_mag + b_mag;
-
-    let grad_den: Vec3 = -(a * s_a + b * s_b + c * s_c);
-
-    let grad_angle: Vec3 = (grad_num * den - grad_den * num) * (2.0 / (num * num + den * den));
-
-    (angle, grad_angle)
+    // Edge potential without regularization
+    ln((s + e) / (s - e))
 }
 
-// Edge potential
-#[allow(non_snake_case)]
-fn edge_potential(P: &Vec3, Q: &Vec3, r: &Vec3) -> f64 {
-    let p: Vec3 = *P - *r;
-    let q: Vec3 = *Q - *r;
-    let e: Vec3 = *Q - *P;
-
-    let p_mag = p.mag();
-    let q_mag = q.mag();
-    let e_mag = e.mag();
-
-    let pq = p_mag + q_mag;
-
-    // Unstable form 
-    // ln((pq + e_mag) / (pq - e_mag))
-
-    // Stable form
-    2.0 * ln(pq + e_mag) - ln(2.0 * (p_mag * q_mag + p.dot(&q)))
-}
-
-// Both edge potential and gradient
-#[allow(non_snake_case)]
-#[inline]
-fn edge_potential_and_gradient(P: &Vec3, Q: &Vec3, r: &Vec3) -> (f64, Vec3) {
-    let p: Vec3 = *P - *r;
-    let q: Vec3 = *Q - *r;
-    let e: Vec3 = *Q - *P;
-
-    let p_mag: f64 = p.mag();
-    let q_mag: f64 = q.mag();
-    let e_mag: f64 = e.mag();
-
-    let pq: f64 = p_mag + q_mag;
-
-    // Edge potential calculation complete
-    // Stable form:
-    let potential: f64 = 2.0 * ln(pq + e_mag) - ln(2.0 * (p_mag * q_mag + p.dot(&q)));
-
-    // Unstable form:
-    // let potential: f64 = ln((pq + e_mag) / (pq - e_mag));
-
-    let gradient: Vec3 = (p / p_mag + q / q_mag) * (e_mag / (p_mag * q_mag + p.dot(&q)));
-
-    (potential, gradient)
-}
-
-// Charge potential of the face
+// Charge potential of the face as defined in eq. 17 in Fabbri (2008)
 #[inline(always)]
 fn charge_potential(face: &Face, r: &Vec3) -> f64 {
-
+    // Normal distance from the plane of the face to the target point
     let d: f64 = face.n_hat.dot(&(*r - face.a));
-    let W: f64 = solid_angle(&face.a, &face.b, &face.c, r);
 
-    let mut result = d * W;
+    // Solid angle subtended by the triangular face
+    let omega: f64 = solid_angle(&face.a, &face.b, &face.c, r);
 
-    let xi: f64 = face.t_hat[0].dot(&(*r - face.a));
-    let gamma: f64 = edge_potential(&face.a, &face.b, r);
-    result += gamma * xi;
+    // Manually unroll the summation over edges so that this function can be
+    // vectorized when inlined by the compiler
+    let mut result: f64 = d * omega;
 
-    let xi: f64 = face.t_hat[1].dot(&(*r - face.b));
-    let gamma: f64 = edge_potential(&face.b, &face.c, r);
-    result += gamma * xi;
-
-    let xi: f64 = face.t_hat[2].dot(&(*r - face.c));
-    let gamma: f64 = edge_potential(&face.c, &face.a, r);
-    result += gamma * xi;
+    // Use the scalar triple product definition to rearrange eq 17
+    // This version uses the in-plane unit vector on the edge, which
+    // saves a cross-product every iteration
+    result += face.t_hat[0].dot(&(*r - face.a)) * edge_potential(&face.a, &face.b, r);
+    result += face.t_hat[1].dot(&(*r - face.b)) * edge_potential(&face.b, &face.c, r);
+    result += face.t_hat[2].dot(&(*r - face.c)) * edge_potential(&face.c, &face.a, r);
 
     result
 }
@@ -624,28 +140,102 @@ fn charge_potential(face: &Face, r: &Vec3) -> f64 {
 // Charge potential gradient from Byzov 2022 (equation 5)
 #[inline(always)]
 fn charge_potential_gradient(face: &Face, r: &Vec3) -> Vec3 {
-    let W: f64 = solid_angle(&face.a, &face.b, &face.c, r);
+    let omega: f64 = solid_angle(&face.a, &face.b, &face.c, r);
 
-    let mut result: Vec3 = face.n_hat * W;
+    let mut result: Vec3 = face.n_hat * omega;
 
     // Vi(a) = 2 atanh(e / S)
-    // e = vector from start node to end node on edge 
-    // S = sum of the magnitudes of the vectors from a to each node 
+    // e = vector from start node to end node on edge
+    // S = sum of the magnitudes of the vectors from a to each node
     let v0: f64 = edge_potential(&face.a, &face.b, r);
     let v1: f64 = edge_potential(&face.b, &face.c, r);
     let v2: f64 = edge_potential(&face.c, &face.a, r);
 
-    // Edge sum 
-    let mut edge_sum: Vec3 = face.e_hat[0] * v0; 
-    edge_sum += face.e_hat[1] * v1; 
-    edge_sum += face.e_hat[2] * v2; 
+    // Edge sum
+    let mut edge_sum: Vec3 = face.e_hat[0] * v0;
+    edge_sum += face.e_hat[1] * v1;
+    edge_sum += face.e_hat[2] * v2;
 
-    result += face.n_hat.cross(&edge_sum); //* 2.0;
+    result += face.n_hat.cross(&edge_sum);
 
     result
 }
 
-/// Solid angle formulation of the above
+/// Compute magnetic field intensity generated by a 4-node tetrahedral finite element
+///
+/// This function operates on n-number of target points and
+/// is vectorized for efficient computation. This version uses the solid angle
+/// formulation (see reference).
+///
+/// # Arguments
+/// - `nodes`: (m), coordinates of each corner node in 3D space
+/// - `jdensity`: (A/m2), current density vector, assumed constant throughout the element
+/// - `targets`: (m), target locations in 3d space
+/// * `out`: (A/m) pre-allocated results accumulation for h-field (hx, hy, hz)
+///
+/// # Note
+/// Accumulates into `out`. Caller must zero `out` before the first call.
+///
+/// # Reference
+/// M. Fabbri, "Magnetic Flux Density and Vector Potential of Uniform Polyhedral
+/// Sources," in IEEE Transactions on Magnetics, vol. 44, no. 1, pp. 32-36, Jan. 2008,
+/// doi: <https://ieeexplore.ieee.org/document/4407584>.
+pub fn h_current_tet4(
+    nodes: &[Vec3; 4],
+    jdensity: &Vec3,
+    targets: (&[f64], &[f64], &[f64]),
+    h: (&mut [f64], &mut [f64], &mut [f64]),
+) {
+    let (x, y, z) = targets;
+    let (hx, hy, hz) = h;
+    let n_targets = x.len();
+    assert_eq!(n_targets, y.len());
+    assert_eq!(n_targets, z.len());
+    assert_eq!(n_targets, hx.len());
+    assert_eq!(n_targets, hy.len());
+    assert_eq!(n_targets, hz.len());
+
+    for f in 0..4 {
+        let (na, nb, nc) = face_nodes(f);
+
+        let a: Vec3 = nodes[na];
+        let b: Vec3 = nodes[nb];
+        let c: Vec3 = nodes[nc];
+
+        let face = precompute_face(&a, &b, &c);
+        let prefactor = jdensity.cross(&face.n_hat) * INV_4PI;
+
+        for i in 0..n_targets {
+            let r = Vec3([x[i], y[i], z[i]]);
+            let psi = charge_potential(&face, &r);
+            hx[i] += prefactor[0] * psi;
+            hy[i] += prefactor[1] * psi;
+            hz[i] += prefactor[2] * psi;
+        }
+    }
+}
+
+/// Compute the magnetic field intensity generated by a uniformly magnetized
+/// tetrahedral element.
+///
+/// This function uses the solid-angle formulation, which is faster and more stable
+/// than the edge-integral formulation.
+///
+/// # Arguments
+/// * `nodes`: (m) coordinates of the corner nodes of the source element
+/// * `mvector`: (A/m) magnetization vector, assumed constant over the element
+/// * `targets`: (m) x,y,z coordinates of each target point
+/// * `out`: (A/m) pre-allocated results accumulation for h-field (hx, hy, hz)
+///
+/// # Note
+/// Accumulates into `out`. Caller must zero `out` before the first call.
+///  
+/// # Reference  
+/// Byzov, D., Martyshko, P., & Chernoskutov, A. (2022).
+/// Computationally Effective Modeling of Self-Demagnetization and Magnetic Field
+/// for Bodies of Arbitrary Shape Using Polyhedron Discretization.
+/// Mathematics, 10(10), 1656.
+/// <https://doi.org/10.3390/math10101656>
 pub fn h_mag_tet4(
     nodes: &[Vec3; 4], // source nodes
     mvector: &Vec3,    // source element
@@ -655,107 +245,23 @@ pub fn h_mag_tet4(
     let (x, y, z) = targets;
     let (hx, hy, hz) = out;
     let n_targets: usize = x.len();
-    let scalar: f64 = -1.0 / (4.0 * PI);
 
     for f in 0..4 {
-        let na: usize = NODE_MAP[f][0][0];
-        let nb: usize = NODE_MAP[f][0][1];
-        let nc: usize = NODE_MAP[f][0][2];
+        let (na, nb, nc) = face_nodes(f);
 
-        let A: Vec3 = nodes[na];
-        let B: Vec3 = nodes[nb];
-        let C: Vec3 = nodes[nc];
-        let face: Face = precompute_face(&A, &B, &C);
+        let a: Vec3 = nodes[na];
+        let b: Vec3 = nodes[nb];
+        let c: Vec3 = nodes[nc];
+        let face: Face = precompute_face(&a, &b, &c);
         let sigma: f64 = mvector.dot(&face.n_hat);
 
         for i in 0..n_targets {
             let r: Vec3 = Vec3([x[i], y[i], z[i]]);
             let grad_psi: Vec3 = charge_potential_gradient(&face, &r);
-            let contribution: Vec3 = grad_psi * (scalar * sigma);
+            let contribution: Vec3 = grad_psi * (-INV_4PI * sigma);
             hx[i] += contribution[0];
             hy[i] += contribution[1];
             hz[i] += contribution[2];
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    // Check that the edge-based current source formulation matches the 
-    // solid angle formulation. Note that they differ in the far field, which
-    // is an area of investigation.
-    #[test]
-    #[should_panic]
-    fn h_current_edge_matches_solid_angle() {
-
-        // Arbitrary tet4 element 
-        let nodes: [Vec3; 4] = [
-            Vec3([0.0, 0.0, 0.0]),
-            Vec3([1.0, 0.0, 0.0]),
-            Vec3([0.0, 1.0, 0.0]),
-            Vec3([0.0, 0.0, 1.0]),
-        ];
-
-        // Arbitrary current vector
-        let j: Vec3 = Vec3([1.0e6, -2.0e4, 7e3]);
-        // let j = Vec3([1.0,0.0,0.0]);
-
-        // Random points in 3D space - near, far, inside element
-        let test_points: [[f64; 3]; 4] = [
-            [3.0, 0.0, 0.0],
-            [0.3, 0.3, 2.0],
-            [10.0, 10.0, 10.0],
-            [-1.0, -1.0, 0.5],
-        ];
-
-        // Workspace for edge-based formulation
-        let mut f: [Vec3; 1] = [Vec3::default(); 1];
-
-        // At every point, check that the fields are equivalent
-        for p in &test_points {
-            let x = [p[0]];
-            let y = [p[1]];
-            let z = [p[2]];
-            let mut hx_edge = [0.0];
-            let mut hy_edge = [0.0];
-            let mut hz_edge = [0.0];
-            let mut hx = [0.0];
-            let mut hy = [0.0];
-            let mut hz = [0.0];
-
-            h_current_tet4(
-                &nodes,
-                &j,
-                (&x, &y, &z),
-                (&mut hx, &mut hy, &mut hz),
-            );
-
-            h_current_tet4_edge(
-                &nodes,
-                &j,
-                (&x, &y, &z),
-                &mut f,
-                (&mut hx_edge, &mut hy_edge, &mut hz_edge),
-            );
-
-            f.fill(Vec3([0.0, 0.0, 0.0]));
-            println!("hx, edge: {} | solidangle: {}", hx_edge[0], hx[0]);
-            println!("hy, edge: {} | solidangle: {}", hy_edge[0], hy[0]);
-            println!("hz, edge: {} | solidangle: {}", hz_edge[0], hz[0]);
-
-            let diff_x: f64 = (hx[0] - hx_edge[0])/(0.5*(hx[0] + hx_edge[0])).abs().max(1e-8);
-            let diff_y: f64 = (hy[0] - hy_edge[0])/(0.5*(hy[0] + hy_edge[0])).abs().max(1e-8);
-            let diff_z: f64 = (hz[0] - hz_edge[0])/(0.5*(hz[0] + hz_edge[0])).abs().max(1e-8);
-
-            println!("Diff x: {} %", 100.0*diff_x.abs());
-            println!("Diff y: {} %", 100.0*diff_y.abs());
-            println!("Diff z: {} %", 100.0*diff_z.abs());
-
-            assert!(diff_x.abs() < 1e-3);
-            assert!(diff_y.abs() < 1e-3);
-            assert!(diff_z.abs() < 1e-3);
         }
     }
 }
